\name{getCor-methods}
\docType{methods}
\alias{getCor-method}
\alias{getCor}
\alias{getCor,ANY-method}
\alias{getCor,RS-method}
\title{Method that extends the cor function of stats to handle reduced segment data}
\description{
  \code{getCor} gets data contained by an RS object ready and then calls the cor function for correlation calculations 
}
\section{Methods}{
\describe{

\item{x = "ANY"}{see function cor of stats}

\item{x = "RS"}{method that handles RS objects}
}}
\keyword{methods}
