\name{diffBy-methods}
\docType{methods}
\alias{diffBy-method}
\alias{diffBy}
\alias{diffBy,ANY-method}
\alias{diffBy,RS-method}
\title{A filter that filters out features that do not differ by a set threshold between a pair of samples}
\description{
  \code{madFilter} With a given threshold, the method checks each pair of samples to exclude features that do not differ by the threshold between two samples. This filter is only applicable to pairs of reduced segments}
\section{Methods}{
\describe{

\item{object = "ANY"}{not iimplemented yet}

\item{object = "RS"}{method that handles RS objects}
}}
\keyword{methods}
