\name{genefilter-methods}
\docType{methods}
\alias{genefilter-methods}
\alias{genefilter,RS-method}
\title{A method that filters feature based on reduced segment}
\description{
    Extends genefilter filtering functions to handle reduced segment data
}
\section{Methods}{
\describe{

\item{expr = "RS"}{A character string to indicate that filtering is based on reduced segment data}
}}
\keyword{methods}
