\name{madFilter-methods}
\docType{methods}
\alias{madFilter-method}
\alias{madFilter}
\alias{madFilter,ANY-method}
\alias{madFilter,RS-method}
\title{Method that filters reduced segment matrix by the mean absolute deviation}
\description{
  \code{madFilter} calculates the mean absolute deviation across samples for each rows and drops rows that are not above percentile defined by a user 
}
\section{Methods}{
\describe{

\item{object = "ANY"}{not iimplemented yet}

\item{object = "RS"}{method that handles RS objects}
}}
\keyword{methods}
